dotted line pattern support
Canvas gets several new calls:

`setLinePattern` accepts a LinePattern object for setting a pattern a line can optionally use

`setLinePatternLength` allows the length of the line pattern to be set

`setLinePatternOffset` allows the offset within the pattern to be adjusted

`setLineOptions` sets whether a line should be drawn as a dotted line, and whether its first and/or last points should be omitted

BitmappedDisplayController’s `PaintState` object now includes `linePattern`, `lineOptions` and `linePatternLength`.

`genericAbsDrawLine` now looks at the `lineOptions` in the current paint state to determine whether a dotted line should be drawn, and whether first or last points should be omitted

fabutils gains a `getBit` function which will get a bit at an index within an array of uint8_t numbers.  the position is in “most significant bit” order, i.e. bitIndex 0 is the bit representing 128 in the first byte in the array